+++ src/ashare_lake/adapters/mofcom/__init__.py
@@ -1,0 +1,5 @@
+"""商务部数据中心 (MOFCOM) adapters."""
+
+from ashare_lake.adapters.mofcom.social_financing import fetch_social_financing
+
+__all__ = ["fetch_social_financing"]